requirements.txt
paho.mqtt.client
Adafruit_DHT